# api.yaml
swagger: "2.0"
info:
  description: "API para cadastro, edição, exclusão e listagem de usuários."
  version: "1.0.0"
  title: "Cadastro de usuários"
host: "localhost:3000"
paths:
  /user:
    post:
      tags:
      - "user"
      summary: "Adiciona um novo usuário"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Usuário a ser cadastrado"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        201:
          description: "Usuário criado"

  /user/{nome}:
    get:
      tags:
      - "user"
      summary: "Recupera um usuário pelo nome"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Sucesso na requisição"
        404:
          description: "Usuário não encontrado"

    delete:
      tags:
      - "user"
      summary: "Exclui um usuário"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Usuário excluído"
        500:
          description: "Usuário não excluído"

    put:
      tags:
      - "user"
      summary: "Atualiza um usuário"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Usuário a ser cadastrado"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        404:
          description: "Usuário não encontrado"
        200:
          description: "Usuário atualizado"
        500:
          description: "Não foi possível atualizar o usuário"

  /users:
    get:
      tags:
      - "user"
      summary: "Lista todos os usuários"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: "Sucesso na requisição"

definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      email:
        type: "string"
      idade:
        type: "integer"
        format: "int64"
    xml:
      name: "User"